requirements.txt
tensorflow
numpy
pandas
matplotlib
seaborn
opencv-python
Pillow
scikit-learn
streamlit
plotly